+++ client/API/index.js
@@ -1,11 +1,9 @@
 import AUTH from './auth-api';
 import GAMES from './games-api';
 import TOURNAMENTS from './tournaments-api';
-import HOTTOURNAMENTS from './hottournaments-api';
 
 export default {
   AUTH,
   GAMES,
   TOURNAMENTS,
-  HOTTOURNAMENTS,
 };
