+++ client/index.html
@@ -4,7 +4,9 @@
     <!--
     <link rel="stylesheet" href="/static/style.css">
     -->
+    <!--
     <script src="https://use.fontawesome.com/b4bc1b634b.js"></script>
+    -->
   </head>
   <body>
     <div id="app"></div>
