+++ client/sagas/tournaments-sagas.js
@@ -13,7 +13,7 @@ function* fetchTournaments() {
 
 function* fetchHotTournaments() {
   yield take(actionTypes.HOT_TOURNAMENT_FETCH);
-  const response = yield call(API.HOTTOURNAMENTS.fetchHotTournament); // todo handle failed responses.
+  const response = yield call(API.TOURNAMENTS.fetchHotTournament); // todo handle failed responses.
   yield put({ type: actionTypes.FETCH_HOTTOURNAMENT_SUCCESS, data: response });
 }
