+++ client/style/_bootstrap.scss
@@ -37,7 +37,7 @@
 //@import "bootstrap/jumbotron";
 //@import "bootstrap/thumbnails";
 //@import "bootstrap/alerts";
-@import "bootstrap/progress-bars";
+//@import "bootstrap/progress-bars";
 //@import "bootstrap/media";
 @import "bootstrap/list-group";
 @import "bootstrap/panels";
