+++ server/routes/index.js
@@ -1,24 +1,17 @@
 import express from 'express';
 import { Render } from '../controllers/index.controller';
 import authRouter from './auth.router';
-import { tournamentRouter, nonvalidTournamentRouter } from './api/tournament.router';
-
-
+import tournamentRouter from './api/tournament.router';
 import gameRouter from './api/game.router';
-
 
 const router = express.Router();
 
 router.get('/', Render);
 router.use('/auth', authRouter);
 
-
-
 // REST APIs
-router.use('/api/tournaments', nonvalidTournamentRouter, tournamentRouter);
-
+router.use('/api/tournaments', tournamentRouter);
 router.use('/api/games', gameRouter);
-router.use('/api/tournaments/hot-tournament', nonvalidTournamentRouter);
 router.get('*', Render);
 
 export default router;
